+++ examples/stream-parser/src/main.rs
@@ -12,6 +12,11 @@ use std::path::PathBuf;
 use clap::Parser as _;
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 use yellowstone_vixen::{self as vixen, proto::parser, vixen_core::proto::Proto};
+use yellowstone_vixen_jupiter_swap_parser::{
+    accounts_parser::AccountParser as JupiterSwapAccParser,
+    instructions_parser::InstructionParser as JupiterSwapIxParser,
+    proto_def::DESCRIPTOR_SET as JUPITER_SWAP_DESCRIPTOR_SET,
+};
 use yellowstone_vixen_meteora_parser::{
     accounts_parser::AccountParser as MeteoraAccParser,
     instructions_parser::InstructionParser as MeteoraIxParser,
@@ -58,18 +63,21 @@ fn main() {
         .descriptor_set(parser::token_extensions::DESCRIPTOR_SET)
         .descriptor_set(METEORA_DESCRIPTOR_SET)
         .descriptor_set(PUMP_DESCRIPTOR_SET)
+        .descriptor_set(JUPITER_SWAP_DESCRIPTOR_SET)
         .account(Proto::new(MeteoraAccParser))
         .account(Proto::new(PumpfunAccParser))
         .account(Proto::new(TokenExtensionProgramAccParser))
         .account(Proto::new(TokenProgramAccParser))
         .account(Proto::new(OrcaAccParser))
         .account(Proto::new(RaydiumAccParser))
+        .account(Proto::new(JupiterSwapAccParser))
         .instruction(Proto::new(MeteoraIxParser))
         .instruction(Proto::new(PumpfunIxParser))
         .instruction(Proto::new(TokenProgramIxParser))
         .instruction(Proto::new(TokenExtensionProgramIxParser))
         .instruction(Proto::new(OrcaIxParser))
         .instruction(Proto::new(RaydiumIxParser))
+        .instruction(Proto::new(JupiterSwapIxParser))
         .build(config)
         .run();
 }